Memory debugging
Squashed a few memory leaks

+++ src/ECS/RenderComponent.h
@@ -25,7 +25,7 @@ struct RenderComponent
     unsigned int ibo;
     unsigned int shader;
     unsigned int texture;
-    unsigned short int vbo;
+    unsigned int vbo;
 };
 typedef struct RenderComponent RenderComponent;
